Feature/video menu (#117)
* #104 feature: draggable frame and component buttons

* #104 refactor: split videopicker and videofeed

* #55 refactor: make VideoFeed functional

* #104 style: style menu

* #104 style: finishing touches
