+++ src/VideoComponents/VideoApp.jsx
@@ -10,6 +10,8 @@ import ModeWidget from './ModeWidget';
 import MiniMapWidget from './MiniMapWidget';
 import GamepadWrapper from './GamepadWrapper';
 import KeyboardWrapper from './KeyboardWrapper';
+import MenuButton from './MenuButton';
+import VideoPicker from './VideoPicker';
 
 const { remote } = window.require('electron');
 
@@ -18,6 +20,7 @@ function VideoApp() {
   const [sensorValues, sensorUpdate] = useState(remote.getGlobal('fromROV'));
   const [biasValues, biasUpdate] = useState(remote.getGlobal('bias'));
   const [transparent, toggleTransparent] = useState(false);
+  const [deviceId, setDeviceId] = useState('');
 
   const closeVideoWindow = () => {
     let w = remote.getCurrentWindow();
@@ -34,6 +37,25 @@ function VideoApp() {
 
   return (
     <div className={transparent ? 'transparentVideoApp' : 'VideoApp'}>
+      <div className="videoMenu">
+        <MenuButton
+          clickFunction={() => toggleTransparent(!transparent)}
+          image="transparent"
+          additionalClass="transparentBtn"
+        />
+        <VideoPicker
+          deviceId={deviceId}
+          handleClick={id => setDeviceId(id)}
+          hidden={transparent}
+        />
+        <MenuButton
+          clickFunction={() => {
+            closeVideoWindow();
+          }}
+          image="close"
+          additionalClass="closeBtn"
+        ></MenuButton>
+      </div>
       <BiasWidget
         u={biasValues['surge']}
         v={biasValues['sway']}
@@ -62,19 +84,7 @@ function VideoApp() {
       />
       <GamepadWrapper className="GamepadWrapper" />
       <KeyboardWrapper className="KeyboardInput" />
-      <VideoFeed hidden={transparent} />
-      <button
-        className="toggleTransparentBtn"
-        onClick={() => {
-          toggleTransparent(!transparent);
-        }}
-      ></button>
-      <button
-        className="closeVideoBtn"
-        onClick={() => {
-          closeVideoWindow();
-        }}
-      ></button>
+      <VideoFeed deviceId={deviceId} hidden={transparent} />
     </div>
   );
 }
