+++ src/VideoComponents/css/LockWidget.css
@@ -21,10 +21,10 @@
   left: -9999px;
   margin: auto;
 
-  top: 125px;
+  top: 135px;
 }
 
 #LockWidgetDepth {
-  left: 90px;
+  left: 87px;
   top: 370px;
 }
