+++ src/VideoComponents/css/VideoPicker.css
@@ -1,0 +1,79 @@
+/* The container <div> - needed to position the dropdown content */
+.dropdown {
+  position: absolute;
+  z-index: 1;
+  display: inline-block;
+  top: -2px;
+  left: 34px;
+}
+
+/* Dropdown Button */
+.dropbtn {
+  padding: 10px;
+  border: none;
+  background-color: transparent;
+  background-image: url('../images/camera.png');
+  background-repeat: no-repeat;
+  background-size: 20px;
+  outline: none;
+  -webkit-app-region: no-drag;
+}
+
+/* Dropdown Content (Hidden by Default) */
+#video-dropdown-content {
+  display: none;
+  position: absolute;
+  background-color: lightgray;
+  z-index: 1;
+  padding: 20px;
+  border-radius: 0 0 5px 0;
+  border: 1px solid rgb(66, 66, 66);
+  border-width: 0 1px 1px 0;
+}
+
+/* Links inside the dropdown */
+#video-dropdown-content button {
+  color: white;
+  background-color: rgb(68, 68, 68);
+  box-shadow: 0 0 1px 1px rgb(43, 43, 43);
+  border-radius: 5px;
+  padding: 12px;
+  text-decoration: none;
+  display: block;
+  width: 200px;
+  border: none;
+  border-width: 1px 0 0 0;
+  cursor: pointer;
+  text-align: left;
+  outline: none;
+  line-height: 20px;
+}
+
+#video-dropdown-content button:not(:first-child) {
+  margin-top: 10px;
+}
+
+/* Change color of dropdown links on hover */
+#video-dropdown-content button:hover {
+  background-color: rgb(51, 51, 51);
+}
+
+/* Show the dropdown menu on hover */
+.dropdown:hover #video-dropdown-content {
+  display: block;
+}
+
+/* Change the background color of the dropdown button when the dropdown content is shown */
+.dropdown:hover .dropbtn {
+  background-color: lightgray;
+  width: 241px;
+  border: 1px solid rgb(66, 66, 66);
+  border-width: 0 1px 0 0;
+}
+
+.selectedFeed::after {
+  content: '\2713';
+  float: right;
+  font-weight: bold;
+  font-size: 20px;
+}
